WIP: Expose Ironic through LoadBalancer instead of host network
This adds Metallb, disables host networking and adds a LoadBalancer
Service. Curling the LoadBalancer IP works great. However, Ironic and
Inspector tries to be smart about what IP they use to communicate and
ends up using the IP of the Pod in the cluster network. The certificate
obviously doesn't have this IP in it so things break down.